Speed up Travis builds by setting osx_image to xcode6.4
The osx_image choice turns out to have a HUGE impact on how long the OSX
Travis builds take. I'm not sure why. Total of all 5 OSX builds in our
matrix are:
    xcode6.4         93 min
    xcode7 (7.0)    120
    xcode7.2        146
    xcode7.3        102

We were (until this patch) using 7.2. This is the worst choice of all,
and is dramatically slower than back when we used the default (which is
xcode 6.1, I think). I don't really want one that old.

For now, I'm sticking with xcode6.4. That's newer than the default (roughly,
clang 3.6). Using 7.3 is also not a bad choice, but for now I'm choosing
raw speed, since on my personal laptop I build against new compilers
so I have that base covered. We will continue to test as Travis makes new
xcode images available, and will upgrade when performance improves. But
it surely seems wise to avoid 7.0 and 7.2.