Rethink default channel naming, just use R, G, B, A (Y for single channel) (#1434)
1-channel images (from formats without channel naming) get Y.
2+ channel images get R, G, B, A, channel4, channel5, ...

Before, we were setting the default names to "A" for 1-channel and
"I","A" for 2-channel.  Which is bad because "A" means alpha, and some
operations treat alpha as special (such as PNG output dividing all the
non-alpha channels by the alpha value!).

This is just for formats that don't actually store channel names
in the files. Which is most formats.