PNG: Better alpha sanity checking and channel names.
The PNG spec is quite particular that a 2-channel image is gray+alpha,
so make sure that we name channels Y,A and designate channel 1 as alpha.

Also, since writing PNG involves automatically un-premultiplying the
color channels by alpha, enforce the fact that for PNG, alpha can ONLY
be the last channel of a 2 or 4 channel image. This prevents an app that
has set channel names or spec.alpha_channel stupidly from messing up the
values in the file (we wouldn't want to un-premultiply by a channel that
can't, by definition, be alpha).