OpenEXR write -- watch out for duplicate channel names.
MOST of the time, the channel names in an ImageSpec are merely
"informational." But one of those times when it makes a real difference
is writing OpenEXR files -- if there are multiple channels with the same
name, libIlmImf will silently DROP the duplicates! And it's not too hard
if you shuffle or replicate channels and aren't careful with the naming,
to do this. So this patch checks the exr output channel names, and
rewrites any duplicate ones to "channelN" (N is the index). It's
probably not "right", but it's probably better than channels
disappearing from the output entirely; it's easier to change a channel
name later than to discover that you've lost all the pixels for a
channel.