oiiotool: improved handling of multi-subimage
Make --origin, --croptofull, and --trim do their thing to all subimages,
if the image at the top of stack has multiple subimages.

Make --crop operate on all subimages if either the -a flag was used, or
if the specific optional override --crop:allsubimages=1 is used.

Change the operation of the generic operators so that the majority of
functions accept a ":allsubimages=..." option that can override the
-a on a command-by-command basis.