More cleanup of bogus channel name edge cases (#1435)
* PNG: Better alpha sanity checking and channel names.

The PNG spec is quite particular that a 2-channel image is gray+alpha,
so make sure that we name channels Y,A and designate channel 1 as alpha.

Also, since writing PNG involves automatically un-premultiplying the
color channels by alpha, enforce the fact that for PNG, alpha can ONLY
be the last channel of a 2 or 4 channel image. This prevents an app that
has set channel names or spec.alpha_channel stupidly from messing up the
values in the file (we wouldn't want to un-premultiply by a channel that
can't, by definition, be alpha).

* OpenEXR write -- watch out for duplicate channel names.

MOST of the time, the channel names in an ImageSpec are merely
"informational." But one of those times when it makes a real difference
is writing OpenEXR files -- if there are multiple channels with the same
name, libIlmImf will silently DROP the duplicates! And it's not too hard
if you shuffle or replicate channels and aren't careful with the naming,
to do this. So this patch checks the exr output channel names, and
rewrites any duplicate ones to "channelN" (N is the index). It's
probably not "right", but it's probably better than channels
disappearing from the output entirely; it's easier to change a channel
name later than to discover that you've lost all the pixels for a
channel.