Change plain text docs to markdown.
README.rst (why RST?) and the plain text CHANGES, LICENSE, CREDITS,
and INSTALL are all converted to MarkDown for nice looking rendering when
browsing the code base on GitHub. The beauty of MarkDown is, it's still
just as readable as ever when viewed as plain text in a terminal window
or an editor.