oiiotool --iconfig NAME VALUE lets you set input configurations.
Very occasionally, you want to do an "open with config" where a
configuration ImageSpec is supplied to the ImageInput::open() and
contains metadata that gives special instructions (invariably they are
format-specific).

It was awkward to do so for images read by oiiotool.

This command lets you set config metadata that will apply to the NEXT
image read by referencing its name on the command line (after which the
configuration metadata will be cleared).

Example:

    oiiotool -iconfig "oiio:UnassociatedAlpha" 1 input.png -o out.exr

    # The PNG input recognizes oiio:UnassociatedAlpha configuration
    # metadata, when nonzero, to cause it to NOT leave PNG's unassociated
    # (not "pre-multiplied" by alpha) color values intact, rather than
    # applying the automatic conversion to associated alpha.