SPI: Make it possible to build a C++03 variant on rhel7.
This is in case we need to link with apps that are still forced to build
pre-C++11. I did this thinking it was responsible for some crashes. Turns
out that wasn't the culprit, but since there's a distinct possibility
that we will discover a need this capability, I want to keep the changes.