Refactor simd.h to separate declarations from implementations.
This doesn't introduce any new functionality, just moves code around
within simd.h MASSIVELY. The net effect is to reorganize the header file
so that it has the important class and functions declared and documented
very cleanly in the first section of the file, and then all the gory
details and architecture-specific implementation following that.  This
makes it so that users of the simd.h functionality can more easily read
the relevant parts of the header to understand the classes, and never
need to poke around the ugly implementation.

We could, in the future, move the implementation details to one or more
secondary files included from here, perhaps split by class or by
architecture.  No need for that now, but this refactor makes it easier.

It also makes it easier to block-copy-and-paste the existing classes to
make new ones, such as I'm planning to do shortly to make 8-wide and
16-wide versions.