ImageCache/TextureSystem: collect stats on total size on disk.
The size statistics reported by IC/TS are the *data* size -- that is,
the size of all pixels of the image if they were read into memory, as
represented in the cache (uncompressed, and in one of the data types
supported by the cache).

But this doesn't accurately reflect the size of the images as they exist
on disk. The three main ways they differ are: (a) the image data doesn't
count the size of the headers, metadata, etc; (b) the pixel data on disk
is often compressed, making it much smaller stored on disk than the
in-cache size; (c) the data on disk may be in a format not directly
supported by the cache, and is "expanded" upon being read (for example,
10 bit per channel data would be expanded to 16 bits per channel in
the cache, palette images might be represented by one byte per pixel
on disk but 3 bytes of RGB per pixel in the cache, and so on).

This patch records the actual file size on disk for each file referenced
by the ImageCache.

Now the stats try to more clearly describe these two separate values:

    Total pixel data size of all images referenced : 120.0 MB
    Total actual file size of all images referenced : 6.6 MB

The stats also can be accessed through the IC or TS APIs, as totals for
all files via getattribute queries for "stat:image_size" and
"stat:file_size", and for individual files using get_image_info queries
of the same name (all are int64 values).