Massive overhaul of simd.h leading to full support of AVX and AVX2. (1484)
The quick summary is that it adds float8, int8, bool8 classes that
are analogous to the 4-way versions. The existing float4, int4, bool4
classes are more or less the way they were before.

So much changed and moved around within simd.h, it's not worth bothering
to understand the diff. Read the new file if you care about the details.

They are quite speedy on AVX/AVX2, but fall back on a mixture of SSE
(4-wide) and scalar if AVX is not available. Note that the float classes
are generally fully performant on AVX, but it takes AVX2 for most of the
8-wide instructions for integers. In many cases, we get pretty close by
doing it in 4-way chunks rather than falling back to fully scalar. If
there are particular methods that need to be faster on AVX 1.0 (or SSE),
I'm happy to give them a closer look, but I'm going forward under the
assumption that the people who care most about performance are on, or
soon moving to, AVX2.

Renamed "mask4" to "bool4" to avoid confusion with "masks" that will be
coming soon when we add support for AVX-512. The old "mask4" is typedefed
to be bool4 for now, for compatibility.

The simd_test is also massively overhauled, not just to test the new
classes, but also significantly beefing up tests of the 4-way SIMD
classes, and doing time trials of nearly every bit of functionality.