Smarter channel_append channel renaming: use subimage name (#1498)
For OpenEXR, channel names are not merely informational -- channel names
*must* be unique within an image part, or else the ones with duplicate
names will be dropped silently when the file is outout.

IBA::channel_append (and therefore `oiiotool --chappend`) tried to protect
against that by renaming duplicate channels "channelN". (As would happen
if you chappend'ed two images that each were R,G,B.)

There's a special case to consider: taking a multi-subimage (or
multi-part, in EXR lingo) and turning into a many-channel image. Like
you'd do with this command:

    oiiotool manyparts.exr -sisplit -chappend -o manychannels.exr

Having channels named R, G, B, channel3, channel4, ... isn't the best
we can do.

This patch modifies the renaming of duplicate channel names within
channel_append so that if the image had a "oiio:subimagename" metadata,
it uses that as a prefix to try to create a unique name.

So, as an example, suppose manyparts.exr had two parts, both with
channels named R, G, and B, and the second part of the image had the
name "specular", then the image resulting from the above command would
end up with channels named R, G, B, specular.R, specular.G, specular.B.

This is likely to be a lot more useful to people. If this *still*
results in a duplicate channel name somehow, then it will fall back on
the old strategy of just calling it "channelN".

While I was messing with that part of the code, I also lifted the
restriction on IBA::channel_append that required the destination image
to be float. Now, although the two input images must match data type,
the output image can be any desired type.