UDIM get_image_info query for "channels" should succeed.
In general, get_image_info (or the texture equivalent, get_texture_info)
fails for UDIM texture patterns, for all but the "UDIM" query (which means,
"is this a UDIM file").

This is because the UDIM name is a pattern, not a concrete file. The
specific file is not resolved until the texture() call itself, because
the actual filename will depend on the st coordinates being looked up.
The get_texture_info query is not for a particular coordinate.

But let's say that you are writing a shader and want different behavior
or code paths depending on properties of the file, such as the number
of channels. Those queries would fail.

So the purpose of this patch is to make the "channels" query to succeed.
It will find the first file to match the pattern (the 0,0 corner may not
exist -- the UDIM set is allowed to be sparse) and query that.  The big
caveat is, this won't be helpful if the textures in the UDIM collection
don't all have the same number of channels. So watch out for that if you
want to throw caution to the wind and do queries on UDIM patterns.

To start with, since this is so inherently dangerous, "channels" is the
only query we support this way. We may expand this set later, if there
are other queries that (a) especially useful, and (b) make sense to
constrain (or assume) to be identical for all individual textures in
the same UDIM set.