After excluding boost < 1.50, we can clean up filesystem.h
We no longer support any boost versions prior to filesystem v3, so we can
simplify a lot of #if guards.

Also change all the many references to boost::filesystem to just
filesystem, and use a 'using namespace filesystem = boost::filesystem'
in a single place at the front. This will make changes minimal in the
future when we switch to experimental::filesystem (newer compilers) or
std::filesystem (coming in C++17).