Make temporary texture filename unique to avoid classes on shared drives. (#1525)
maketx (and IBA::make_texture underneath) writes its results to a
temporary file to avoid overwriting an existing file and leaving an
invalid file there if the command should be terminated before it's done.
After it's done being written, the temp file is "moved" to the final
destination, replacing it atomically.

But as Ramon points out, if multiple processes are running on the same
shared disk partition and might try to maketx the same texture at the
same time, they could clobber each other by both writing to identically
named temp files. So the solution is to ask the OS to make a unique name
for that temporary file, using Filesystem::unique_path().