Be more careful about building iv man page (#1514)
On Unix systems if txt2man is found, we automatically build man pages
from our command-line binaries. It was smart enough to skip doing so
for iv if USE_QT was 0 (that is, if the build was set up to purposely
exclude anything using QT). But if USE_QT was not set to 0 but iv was
skipped anyway (perhaps Qt or other iv-specific dependencies weren't
found at build time), it would fail in trying to make the man page
for iv, since the iv program itself could not be executed.

This logic fix instead predicates it on whether iv actually got built,
regardless of whether its failure to build was due to explicitly
skipping it, dependencies not found, or whetever.