nuke: Fix txReader to properly restore saved mip level knob value
The user-facing Enumeration_knob is populated from within the Reader's
constructor, which is too late when the Read is created with a non-default mip
level (e.g. when loading a saved script).

To work around this without resorting to saving all of the knob's menu items,
it is no longer stored by the Read or used in the op hash. Instead, its value
is copied to a hidden Int_knob, which drives the actual Reader and gets saved
when the Read is serialized.

Also, the txReader now sets the bit-depth metadata key properly when the input
file contains half data, which was an earlier oversight.