oiiotool -i:type=...
This new option to `-i` forces a read of the input file into a particular
data type. It works a little like --native, except that you can specify
the type explicitly, and on a per-file basis.

Just to review, the full logic of what types oiiotool stores internally
in its ImageBuf's is:

* If `-i:type=...` is used, then it will force a read into an ImageBuf
  of that type.

* Otherwise, if `--native` is used, it will try to store internally in
  the "native" data type of the file, which is typically going to be
  the "highest precision" type of any channel found in the file.

* Otherwise, it will internally convert to FLOAT, which generally speaking
  is going to be the most efficient choice in terms of math performance.

So why use anything but the auto-conversion to float? If you are using
oiiotool merely for file format conversions, data format conversions, or
channel shuffling (i.e, no "math", for which float performance will be
better), you may prefer sticking with the native format or forcing a
format in order to yield better memory usage or to avoid any loss of
precision in the round trip conversion.