Fix warnings on some 32 bit platforms (#1539)
For some reason, certain 32 bit platforms have gcc errors about this
spot leading to uninitialized variables (they don't! but gcc can't seem
to figure that out). The 64 bit platforms figure it out just fine, so
just eat the cost of a couple useless assigns on the platforms where
it's a problem.