Fix errors when doing identify transform when no OCIO is present. (1551)
Well, sometimes you (and by "you" I mean some script you wrote) ask
oiiotool to do a 'colortransform sRGB sRGB' or some other color
transform that is just identity. Fine. But if $OCIO environment variable
is not defined so no OCIO config is found, that was an error!

Simple fix -- detect that case and return a do-nothing color processor.

Fixes #1550