Color maps
* ImageBufAlgo::color_map (C++ and Python)
* oiiotool --colormap

Both take the intensity of each pixel in the input image (either one
channel, or the luminance of its RGB channels) and run it through a
linearly interpolated color map. Built in color maps that can be
specified by name include "blue-red", "spectrum", and "heat". Or raw
values can be used.