oiiotool: New expression substitution: MINCOLOR, MAXCOLOR, AVGCOLOR
In the expression substitution context, these "pseudo-metadata" names
will evaluate to be comma-separated lists of the per-channel pixel value
minimum, maximum, and average, respectively.

As an example, the following oiiotool command line will do contrast
stretching, remapping intensity values to fill the full range of [0-1]:

    oiiotool in.tif -subc {TOP.MINCOLOR} -divc {TOP.MAXCOLOR} -o out.tif

It performs this by first subtracting from every pixel the minimum pixel
value of the input (at that point, the TOP image on the stack), and then
dividing every pixel by the new maximum value, resulting in an output
image whose minimum pixel value is (0,0,0) and maximum is (1.0,1.0,1.0).