thread.h: modernize, thread_pool, parallel_for
* Bring thread_group into modern C++11 era: use std components, no more
  boost dependency, use variadic templates for a better create_thread
  signature. Also remove the last vestigages of using boost::thread_group
  instead of ours.

* Modernize spin_mutex to use C++11 std::atomic_flag instead of doing
  everything with bespoke code.

* thread_pool : persistent threads feeding from a task queue. This is
  generally a better way forward than explicitly spawning new threads
  with thread_group every time you want to do things in parallel.
  OIIO attribute("nthreads") sets the size of the thread pool.

* task_set : Because the thread_pool is asynchronous (fire and forget,
  or fire and get future's back), a simple task_set object lets you collect
  a set of related asynchronous tasks and wait for them to all have been
  completed before proceeding.

* parallel_for : several easy constructs for a variety of parallel
  loops, using the default thread pool. Makes parallelizing "1-D" loops
  very similar to writing an ordinary 'for' loop, or calling
  std::for_each. These are in a new header file, parallel.h.

The implementation of the thread pool is based on the CTPL project
(Apache license). It's hidden completely in thread.cpp, totally
independent of the thread.h exposed APIs that use the PIMPL idiom, so we
are free to replace the underlying pool implementation if we realize
with more experience that CTPL was not the best choice.

Note also that we have bumped the minimum version of Boost to 1.53 to
accommodate the thread_pool's internal use of boost::lockfree::queue.
That seems reasonable, considering that VFX recommended Boost 1.53 as
the recommended version in 2014, 1.55 in 2015 & 2016. Boost lockfree
queue is header-only, so this doesn't introduce any new linkage or
version mismatch issues.

Read the comments in thread.h for a full explanation, but just for
illustration in this review, here is what it looks like to use it:

Serial tasks:

    for (int i = 0; i < N; ++i)
        job(i);

Parallel tasks using thread_pool, and waiting for them to finish before
proceeding:

    thread_pool *p = thread_pool::get_thread_pool();
    work_set w;
    for (int i = 0; i < N; ++i)
        w.push (p->push (job, i));
    w.wait();

Serial loop:

    for (int i = 0; i < N; ++i) {
        a[i] += b[i];
    }

Parallel loop, will divide the range among threads:

    parallel_for (0, N, [&](int64_t i){
        a[i] += b[i];
    });