Fix threadpool
I discovered that the thread_pool could deadlock if the pool queue
filled and then the subtasks themselves added further subtasks to
the queue and waited for their results (since those subtasks would
never themselves get around to executing, with all the pool threads
waiting).

The solution is primarily to modify task_set::wait() so that instead
of unconditionally blocking for the subtasks to complete, it polls
them without blocking, and so when it finds itself waiting for queue
results to still complete, it steals some work from the queue to do
itself. This unclogs the system nicely.

I also modified parallel_for_chunked (which is the basis for the
unchunked parallel_for as well) so that the calling thread itself
executes the last (or only) subtask rather than send it to the pool.
This is yet another way to reduce pressure on the queue, and especially
for cases where you call parallel_for but only one task is necessary
(due to chunking size), it avoids any overhead of using the pool.