oiiotool -fit:exact=1
The optional exact=1 modifier will cause the fit's resize to achieve an
exact match of aspect ratio and centering, to partial-pixel precision if
necessary (but at the risk of blurring slightly, especially at the
edges). The default (exact=0) behavior, as always, will do the resize to
a whole-number pixel resolution, which will achieve better sharpness but
may not exactly preserve the aspect ratio nor exactly center the padded
images (since that may be impossible without partial pixel resolution).