SPI: more correct de-duplication of tiff.imageio target
Strange linkage problems specific to SPI's setup required the tiff
target to be first. I had tried to add it to the beginning of the list
and remove duplicates, but that didn't actually work properly (because
the GLOB gave me an absolute path, and I inserted a relative path, so
they weren't true duplicates. This was asymptomatic until a recent
'ninja' upgrade on my laptop, which spews warnings about duplicate
targets.  So this cleanup fixes it all again.

new ninja gave errors when we allowed duplicate