parallel_for_chunked: avoid divide by 0 when pool size is 0 (#1568)
Also avoid sizing thread to 0 by default (it better be on purpose).

This is an edge case I don't expect to see much in the real world, but
it was really messing with our Travis-CI builds, which seem to run on
VMs that think they have only one core allocated. This would initialize
the pool to 0 size, which could lead to a math exception, ugh.