thread_pool: improved logic for wanting no extra worker threads.
The thread pool size is the number of extra worker threads, not counting
the calling thread (which also participates in the work).

But if you truly want to run single-threaded, with no extra worker
threads, but still use the therad_pool/parallel_for interface, there
wasn't a correct way to do that, since resize(1) would give you one
worker thread + the caller, and resize(0) would use the default number
of cores.

So the change is that resize(-1) means "use hardware_concurrency-1"
and resize(0) means "no worker threads".

Now, when tasks are pushed onto the thread_pool, if there are no
workers, it will just run the task immediately on the calling
thread. This isn't quite as efficient as bypassing the pool entirely,
because it still needs to package up a lambda and make a future. But the
single-threaded case is not really the case we're optimizing for, and at
least this lets you have a uniform calling sequence (or use of
parallel_for) that doesn't break if you have some reason to not want any
extra worker therads at all. For any really expensive task, I don't think
you'll notice the overhead.