Build option: SANITIZE=...
This lets you do a build that utilizes the new sanitizers (for starters,
address and leak sanitizer). This is helpful for automatically flagging
a wide variety of buffer overrun and leak errors.

This was tricky for testsuite tests that use Python, because Python
itself isn't linked against the asan libraries, so crashes when it
imports the OpenImageIO module that needs it. No big dea, we just
exclude python tests when running SANITIZE, and change a few tests that
relied on Python to be pure oiiotool calls.