maketx did not incorporate "highlight compensation" into the hash! (#1599)
The results when comparing "maketx --hicomp" versus without that flag
were especially hilarious when using 'iv' to view the images -- for high
resolution images, it will fall back on ImageCache, which thinks two
images with the same hash are identical and don't need to be redundantly
read from disk. So, the hicomp and nohicomp images looked identical in
iv, leading to even more confusion.

So, also turn off the ImageCache de-duplication for iv so we don't
get bitten by this again. I don't think there is a legit case where we
rely on deduplication for iv, so this seems like the safe choice.

Also amend the testsuite/maketx to check the hashes and that choice
of filter and --hicomp alter the hashes.