Allow raw CMYK read from TIFF. (1605)
The default is (per OIIO conventions) to convert "non-spectral" color
spaces (such as CMYK) to RGB upon input. But we now establish a new
'configuration' setting, wherein the input configuration attribute
"oiio:RawColor" set to nonzero will be a request to return raw pixel
values, regardless of their apparent color space, with no default
conversion to RGB.

This should be honored by all format readers that allow
non-spectral-like color spaces. The other obvious one that comes to mind
is PSD, which already honored an undocumented "psd:RawColor", which this
patch also alters to accept the more generic "oiio:RawColor" and
documents its behavior for both PSD and TIFF.

For those of you using oiiotool, the way you force input of raw CMYK
values is like this:

    oiiotool cmyk_input.tif -o rgb.exr
    oiiotool -iconfig oiio:RawColor 1 cmyk_input.tif -o cmyk.exr

You'll see that the default behavior (first example) will result in
a 3-channel RGB image output to rgb.exr, whereas the second example
where we set the input configuration will output a 4-channel image
with channel names "C", "M", "Y", and "K". You're on your own for how
to display that properly, but it will be the original raw CMYK values.