Trim the Travis build matrix a bit.
Cut out a couple of the OSX builds (debug, and C++14), because the
Travis OSX farm is small and seems to always be saturated, and this is
causing our CI tests to be significantly delayed. Since I build on OSX
daily on my own laptop, I'm reasonably confident that I'll notice right
away if something breaks the build there, so I don't think this
increases risk significantly for the payoff in faster CI build
turnaround. They're just commented out, I'll re-enable it if TravisCI
increases the size of their OSX pool.

On the Linux front, also eliminate the gcc5 build. I think that people
are split between being stuck on gcc4.8 (per VFXPlatform rules), or are
staying relatively current with gcc6, and gcc7 is on the way. So I don't
think there's much of a gcc5 constituency, and very few potential breaks
that would be unique to 5 (and not break 4.8 or 6).

Aside: It would be nice if Travis allowed a two-tiered set of tests, so
that you could fire off a comprehensive "pre-release" build matrix that
was more thorough than the usual matrix you run to verify any random
checkin or PR.