For error conditions, print the full command line. (#1614)
It's common to call oiiotool from a script, which may assemble the
command line programmatically. If it botches this and ends up with a
mangled command line, it really helps in debugging the script if its
author can see exactly what oiiotool got as a command (essentially,
as part of the original error message).