RAW: Support RAW input to ACES color space. (1626)
The "raw" reader by default will convert its input to sRGB color space,
but allows you to request that it be converted upon read to other
color space (or not alter the values) by setting the "configuration"
option "raw:ColorSpace" to the name of the color space you want.

This patch adds "ACES" to the list of color spaces it understands,
at least for the versions of libraw that support it (version 0.18 and
newer, it seems).

While I was in there, I changed the color space names it recognizes
to be case-insensitive.