Improvement to decoding metadata from files with XMP blocks
OK, this is subtle. When decoding XMP, I originally thought that it
wouldn't include Exif data, but sometimes it does. But the XML is just
string data.  So this patch will look up the name to see if it's a known
Exif field, and if so, use the type we know it ought to be to correctly
decode from a string.

Also in exif.cpp, changed the implementation of the tag map used by
exif_tag_lookup so that it's case insensitive (because, ugh, some
raw images have their Exif data in XMP as "exif" rather than "Exif").

The bottom line is that these changes improve proper recognition of
certain metadata when reading RAW files.