Minor bug fixes in raw file input. (1629)
Fixes #1628

1. The default value for missing "raw:use_camera_matrix" attribute should
have been 1, not 0, in order to match the default behavior of libraw/dcraw.

2. The demosicing control logic wanted to erase the "raw:Exposure"
attribute, but incorrectly specified the search to be constrainted to a
string attribute (thereby failing to erase the attribute which is
actually a float). The simple solution is that those erase_attribute
calls really don't need to constrain the type at all.

3. Add support for newer (as of LibRaw 0.16) demosaic algorithms "DMT"
(mode 11) and "AAHD" (mode 12). Also allow the demosaic algorithm names
to be case insensitive.

4. Rename "Modified AHD" to "AHD-Mod", to simplify the name and
match libraw terminology.