TIFF: Change interpretation of XPOSITION,YPOSITION.
The TIFF spec says these (which we use to generate the ImageSpec x,y
fields giving the data window origin of the image) should be relative
to the "resolution units" described by the XRESOLUTION and YRESOLUTION.

But OIIO has always just stored the data origin directly in this field,
without regard to any scaling factor of the [XY]RESOLUTION fields.

Now, let's realize that it's not very common to have files with nonzero
data origin (crop windows) in TIFF files, also not common to have any
RESOLUTION tags in the files. So the problematic combination, which
would be interpreted differently by versions on either side of this
patch, is rare.

But one way that we can get into trouble is the following sequence: (a)
a JPEG from a camera (or any number of other cases) has resolution or
density data (common for cameras); (b) that JPEG is converted to TIFF,
thereby copying the fields; (c) the TIFF file is cropped, and written by
a version of OIIO prior to this patch. Oops.

Mitigation strategy: OIIO (at least insofar as oiiotool and maketx)
likes to store its name and version in the "Software" TIFF field. So
upon input, we check this, and if the file seems to have been written
by OIIO with a version prior to this patch, use the old interpretation.

Fixes #1623